+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 - 调整文件目录结构 | Adjust file directory structure
 - 配置构建选项 | Configure build options
 - 配置 npm 发布选项 | Configure npm release options
+- 添加 API `/api/system/currentTime` | Add API `/api/system/currentTime`